Refactor: Modularize A.C.A.S codebase with new utility modules
- Updated HTML files to remove direct tag manager script references and added comments for clarity.
- Introduced new JavaScript modules for board utilities, configuration management, DOM element selectors, engine utilities, global variables, internationalization, and user security.
- Implemented a centralized toast notification system for user feedback.
- Added functionality for managing user profiles, settings, and dynamic content insertion.
- Enhanced multithreading detection for WASM support in chess engines.
- Established a communication bridge between the A.C.A.S GUI and the userscript for cross-tab messaging.
- Improved input handling and dropdown management utilities for better user experience.
