+++ app/index.html
@@ -32,16 +32,7 @@
 		gtag("config", "GTM-5499BQ4B");
 		</script>
 		<script async src="https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js?client=ca-pub-7248123202489335" crossorigin="anonymous"></script>
-		<script type="module" src="../assets/js/acas-load-modules.js"></script>
-		<script src="../assets/js/acas-globals.js"></script>
-		<script src="../assets/libraries/CommLink/CommLink.js"></script>
-		<script src="../assets/js/acas-userscript-bridge.js"></script>
-		<script src="../assets/js/acas-i18n-processor.js"></script>
-		<script src="../assets/js/acas-board-analyzer.js"></script>
-		<script src="../assets/js/acas-move-eval.js"></script>
-		<script src="../assets/js/acas-backend-instance.js"></script>
-		<script src="../assets/js/acas-backend-manager.js"></script>
-		<script src="../assets/js/acas-user-security-system.js"></script>
+		<script type="module" src="../assets/js/modules/load-modules.js"></script>
 		<script src="../assets/libraries/FileSaver/FileSaver.js"></script>
 		<script src="../assets/libraries/UniversalBoardDrawer/UniversalBoardDrawer.js"></script>
 		<script src="../assets/libraries/SnapDOM/snapdom.min.js"></script>
