--- assets/js/acas-load-modules.js
@@ -1,5 +1,0 @@
-import { Chessground } from '../libraries/chessgroundx/chessground.js';
-import zerofish from '../../app/assets/engines/zerofish/zerofishEngine.js';
-
-window.ChessgroundX = Chessground;
-window.zerofish = zerofish;
