+++ assets/js/modules/index.js
@@ -1,0 +1,19 @@
+// index.js
+// Barrel file for all A.C.A.S modules. Import from here for easier access and scalability.
+
+export * from './config-management.js';
+export * from './board-utils.js';
+export * from './multithread-status.js';
+export * from './profile-management.js';
+export * from './load-modules.js';
+export * from './settings-management.js';
+export * from './i18n-utils.js';
+export * from './site-additional-content.js';
+export * from './tag-manager.js';
+export * from './toast.js';
+export * from './user-security.js';
+export * from './userscript-bridge.js';
+export * from './utilities.js';
+export * from './engine-utils.js';
+export * from './dom-elements.js';
+export * from './globals.js';
