+++ assets/js/modules/site-additional-content.js
@@ -1,11 +1,19 @@
-function insertGas(elem) {
+// site-additional-content.js
+// Handles Lottie animations, ad insertion, and dynamic content for A.C.A.S.
+// Part of the modular ES6 codebase in assets/js/modules/.
+// Used by the GUI for enhanced visuals and monetization.
+
+// Site additional content utilities for A.C.A.S
+
+/**
+ * Insert a Google AdSense ad into the given element.
+ * @param {HTMLElement} elem
+ */
+export function insertGas(elem) {
     const text = elem?.dataset?.t;
     const rect = elem?.dataset?.r;
-
     elem.dataset.processed = true;
-
     const pubId = "ca-pub-7248123202489335";
-
     if(text) {
         elem.innerHTML = `<ins class="adsbygoogle"
         style="display:block"
@@ -26,13 +34,14 @@ function insertGas(elem) {
         data-ad-client="${pubId}"
         data-ad-slot="4819810311"></ins>`;
     }
-
     (adsbygoogle = window.adsbygoogle || []).push({});
 }
 
-function initLottieAndGas() {
+/**
+ * Initialize Lottie animation and Google Ads on the page.
+ */
+export function initLottieAndGas() {
     const lottieElement = document.querySelector("#lottie-animation");
-
     if (lottieElement && !lottieElement.dataset.lottieLoaded) {
         lottie.loadAnimation({
             container: lottieElement,
@@ -41,18 +50,19 @@ function initLottieAndGas() {
             autoplay: true,
             path: '../assets/json/lottie.json'
         });
-        lottieElement.dataset.lottieLoaded = "true"; // Prevent duplicate loading
+        lottieElement.dataset.lottieLoaded = "true";
     }
-
     [...document.querySelectorAll('.gas')]
         .filter(x => !x.dataset.processed)
         .forEach(insertGas);
 }
 
-initLottieAndGas();
-
-const observer = new MutationObserver((mutations, obs) => {
-    initLottieAndGas();
-});
-
-observer.observe(document.body, { childList: true, subtree: true });
+/**
+ * Observe changes in the document and initialize additional content when needed.
+ */
+export function observeAdditionalContent() {
+    const observer = new MutationObserver((mutations, obs) => {
+        initLottieAndGas();
+    });
+    observer.observe(document.body, { childList: true, subtree: true });
+}
