+++ board/index.html
@@ -9,12 +9,11 @@
 		<meta property="og:title" content="Board / A.C.A.S">
 		<meta property="og:description" content="Full‑screen chess board viewer with FEN and orientation params. Supports animated sequences.">
 		<meta name="viewport" content="width=device-width, initial-scale=1.0">
-		<script src="../assets/js/acas-globals.js"></script>
-		<script src="../assets/js/acas-i18n-processor.js"></script>
+		<!-- Removed old monolithic scripts, now using modular loader below -->
 		<script src="../assets/libraries/FileSaver/FileSaver.js"></script>
 		<script src="../assets/libraries/html2canvas/html2canvas.min.js"></script>
 		<script src="../assets/libraries/UniversalBoardDrawer/UniversalBoardDrawer.js"></script>
-        <script type="module" src="../assets/js/acas-load-modules.js"></script>
+		<script type="module" src="../assets/js/modules/load-modules.js"></script>
 		<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.10.5/font/bootstrap-icons.css"/>
 		<link rel="stylesheet" type="text/css" href="../assets/css/chessground.base.css"/>
 		<link rel="stylesheet" type="text/css" href="../assets/css/chessground.brown.css"/>
