+++ index.html
@@ -21,12 +21,12 @@
     <link rel="icon" type="image/x-icon" href="favicon.ico">
     <link rel="canonical" href="https://psyyke.github.io">
     <meta property="og:type" content="website">
-    <script type="text/javascript"src="assets/js/tag-manager.js"></script>
-    <script async="" type="text/javascript" src="https://www.googletagmanager.com/gtm.js?id=GTM-5499BQ4B"></script>
+    <!-- Tag manager now loaded via modules if needed -->
     <script async="" type="text/javascript" src="index.js"></script>
     <script defer="" type="text/javascript" src="assets/libraries/klaro/config.js"></script>
     <script defer="" type="text/javascript" src="assets/libraries/klaro/klaro.js"></script>
     <script async="" src="https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js?client=ca-pub-7248123202489335" crossorigin="anonymous"></script>
+    <script type="module" src="assets/js/modules/load-modules.js"></script>
 </head>
 <body>
     <noscript><iframe src="https://www.googletagmanager.com/ns.html?id=GTM-5499BQ4B" height="0" width="0" style="display:none;visibility:hidden"></iframe></noscript>
