+++ usage/index.html
@@ -20,7 +20,7 @@
     <link rel="icon" type="image/x-icon" href="../favicon.ico">
     <link rel="canonical" href="https://psyyke.github.io/A.C.A.S/usage">
     <meta property="og:type" content="website">
-    <script type="text/javascript"src="../assets/js/tag-manager.js"></script>
+    <!-- Tag manager now loaded via modules if needed -->
     <script src="../assets/libraries/bodymovin/lottie.js" type="text/javascript"></script>
     <script async="" type="text/javascript" src="https://www.googletagmanager.com/gtm.js?id=GTM-5499BQ4B"></script>
     <script async="" type="text/javascript" src="../index.js"></script>
